+++ src/main/scala/com/github/gvolpe/scalar2019/tagless/exp/clunky.scala
@@ -1,13 +1,14 @@
 package com.github.gvolpe.scalar2019.tagless.exp
 
 import com.github.gvolpe.scalar2019.rio._
+import com.github.gvolpe.scalar2019.tagless.instances.mtl._
 import com.olegpy.meow.hierarchy._
 import cumbersome._
 import scalaz.zio._
 import scalaz.zio.interop.catz._
+import scalaz.zio.interop.catz.mtl._
 
 object clunky extends App {
-  import com.github.gvolpe.scalar2019.rio.instances.mtl._
 
   // Natural transformation (~>) replaces `provide` in polymorphic code
   implicit val fk = RIO.functionK(new Graph[Task].appModule)
